refactor voucher request reducer
