+++ ts/features/bonus/siciliaVola/store/reducers/voucherRequest.ts
@@ -1,0 +1,141 @@
+import { none, Option, some } from "fp-ts/lib/Option";
+import {
+  Company,
+  Hospital,
+  University,
+  VoucherRequest
+} from "../../types/SvVoucherRequest";
+import {
+  FlightsDate,
+  svGenerateVoucherSelectCategory,
+  svGenerateVoucherSelectCompany,
+  svGenerateVoucherSelectFlightsDate,
+  svGenerateVoucherSelectHospital,
+  svGenerateVoucherSelectUniversity,
+  svGenerateVoucherStart,
+  svGenerateVoucherUnderThresholdIncome
+} from "../actions/voucherGeneration";
+import { Action } from "../../../../../store/actions/types";
+import { getType } from "typesafe-actions";
+
+const INITIAL_STATE: Option<VoucherRequest> = none;
+
+/**
+ * Utils needed to update only the under threshold value
+ * @param state
+ * @param underThresholdIncome
+ */
+const updateUnderThresholdIncome = (
+  state: Option<VoucherRequest>,
+  underThresholdIncome: boolean
+): Option<VoucherRequest> =>
+  state.fold(state, vR => {
+    if (vR.category === "worker" || vR.category === "sick") {
+      return some({ ...vR, underThresholdIncome });
+    }
+    return state;
+  });
+
+/**
+ * Utils needed to update only the university value
+ * @param state
+ * @param university
+ */
+const updateUniversity = (
+  state: Option<VoucherRequest>,
+  university: University
+): Option<VoucherRequest> =>
+  state.fold(state, vR => {
+    if (vR.category === "student") {
+      return some({ ...vR, university });
+    }
+    return state;
+  });
+
+/**
+ * Utils needed to update only the company value
+ * @param state
+ * @param company
+ */
+const updateCompany = (
+  state: Option<VoucherRequest>,
+  company: Company
+): Option<VoucherRequest> =>
+  state.fold(state, vR => {
+    if (vR.category === "worker") {
+      return some({ ...vR, company });
+    }
+    return state;
+  });
+
+/**
+ * Utils needed to update only the hospital value
+ * @param state
+ * @param hospital
+ */
+const updateHospital = (
+  state: Option<VoucherRequest>,
+  hospital: Hospital
+): Option<VoucherRequest> =>
+  state.fold(state, vR => {
+    if (vR.category === "sick") {
+      return some({ ...vR, hospital });
+    }
+    return state;
+  });
+
+/**
+ * Utils needed to update only the flights date values
+ * @param state
+ * @param flightsDate
+ */
+const updateFlightsDate = (
+  state: Option<VoucherRequest>,
+  flightsDate: FlightsDate
+): Option<VoucherRequest> =>
+  state.fold(state, vR => {
+    if (
+      vR.category === "student" ||
+      vR.category === "worker" ||
+      vR.category === "sick"
+    ) {
+      return some({
+        ...vR,
+        departureDate: flightsDate.departureDate,
+        returnDate: flightsDate.returnDate
+      });
+    }
+    return state;
+  });
+
+const reducer = (
+  state: Option<VoucherRequest> = INITIAL_STATE,
+  action: Action
+): Option<VoucherRequest> => {
+  switch (action.type) {
+    case getType(svGenerateVoucherStart):
+      return INITIAL_STATE;
+    case getType(svGenerateVoucherSelectCategory):
+      return some({
+        category: action.payload
+      });
+
+    case getType(svGenerateVoucherUnderThresholdIncome):
+      return updateUnderThresholdIncome(state, action.payload);
+
+    case getType(svGenerateVoucherSelectCompany):
+      return updateCompany(state, action.payload);
+
+    case getType(svGenerateVoucherSelectHospital):
+      return updateHospital(state, action.payload);
+
+    case getType(svGenerateVoucherSelectUniversity):
+      return updateUniversity(state, action.payload);
+
+    case getType(svGenerateVoucherSelectFlightsDate):
+      return updateFlightsDate(state, action.payload);
+  }
+  return state;
+};
+
+export default reducer;
